services/vlm/requirements.txt
fastapi==0.104.1
requests
sentencepiece
uvicorn[standard]==0.24.0
torch==2.1.0
transformers==4.36.0
accelerate==0.25.0
bitsandbytes==0.41.3
pillow==10.1.0
python-multipart==0.0.6
numpy<2.0.0
protobuf
scipy  